Enable MathJax
The original plan was to use custom-head.html to add MathJax HTML tags,
but since minima 2.5.2 (latest as of Jan, 2025) doesn't support the file,
I decided to copy head.html.

The copied head.html had a change to the path to the scss file
(see https://github.com/jekyll/minima/pull/305 for the details),
so `assets/main.scss` had to be moved to `assets/css/style.scss`.